Generic test runner: Migrate MUI V5 Input (8/n)
[ghstack-poisoned]
