+++ packages/component-driver-mui-v5-test/__tests__/Input.dom.test.ts
@@ -1,0 +1,10 @@
+import { jestTestAdapter } from '@atomic-testing/jest';
+import { createTestEngine } from '@atomic-testing/react';
+import { testRunner } from '@atomic-testing/test-runner';
+import { basicInputExample, basicInputTestSuite } from '../src/examples';
+
+testRunner(basicInputTestSuite, jestTestAdapter, {
+  getTestEngine: (scenePart: typeof basicInputExample.scene) => {
+    return createTestEngine(basicInputExample.ui, scenePart);
+  },
+});
