+++ packages/component-driver-mui-v5-test/__tests__/Input.e2e.test.ts
@@ -1,0 +1,5 @@
+import { getTestRunnerInterface, playWrightTestFrameworkMapper } from '@atomic-testing/playwright';
+import { testRunner } from '@atomic-testing/test-runner';
+import { basicInputTestSuite } from '../src/examples';
+
+testRunner(basicInputTestSuite, playWrightTestFrameworkMapper, getTestRunnerInterface());
